requirements.txt
albumentations==1.3.1
fastapi==0.112.1
Flask==3.0.3
GitPython==3.1.41
GitPython==3.1.43
ipython==8.12.3
matplotlib==3.8.2
mss==9.0.1
numpy==2.0.1
onnx==1.15.0
onnxruntime==1.17.3
onnxruntime_gpu==1.17.1
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.9.0.80
pandas==2.2.2
Pillow==10.4.0
psutil==5.9.8
psutil==5.9.0
pybboxes==0.1.6
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
scipy==1.14.0
seaborn==0.13.2
setuptools==72.2.0
streamlit==1.37.1
tensorrt==8.6.1.post1
thop==0.1.1.post2209072238
tqdm==4.66.1
ultralytics==8.2.77
ultralytics_thop==2.0.0
uvicorn==0.30.6
